Add ability to set configuration over usb.

+++ packet.h
@@ -7,8 +7,8 @@ class Packet {
     int GetType() {return m_type;}
     void AddTag(uint8_t tag, uint8_t buflen, char *buf);
 
-    bool ReadTag(uint8_t tagnum, uint8_t *value);
-    bool ReadTag(uint8_t tagnum, uint8_t **value);
+    bool ReadTag(uint8_t tagnum, uint16_t *value);
+    bool ReadTag(uint8_t tagnum, uint16_t **value);
 
     uint8_t* FindTag(uint8_t tagnum);
     void AppendBytes(char *buf, int buflen);
